Move client/{tx,rest,utils} into x/auth/client
Ref: #4488
