--- client/rest/errors.go
@@ -1,5 +1,0 @@
-package rest
-
-import "errors"
-
-var errInvalidGasAdjustment = errors.New("invalid gas adjustment")
